Create GlobalVariableUpdateNiFi
This is the groovy script to update global variable in NiFi. I deployed this on my local NiFi so no Bearer token required in this case. NiFi Rest API is used here to update it. You need to get the value from flowfile attributes that you want to update in Global Variables and post it in NiFi Rest API payload as above.